Initial reimplementation of account selector
